Exception handling

+++ src/main/java/com/lgo/dscommerce/services/ProductService.java
@@ -3,10 +3,16 @@ package com.lgo.dscommerce.services;
 import com.lgo.dscommerce.dto.ProductDTO;
 import com.lgo.dscommerce.entities.Product;
 import com.lgo.dscommerce.repositories.ProductRepository;
+import com.lgo.dscommerce.services.exceptions.DatabaseException;
+import com.lgo.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
+import org.hibernate.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -22,16 +28,9 @@ public class ProductService {
 
     @Transactional(readOnly = true) //somente para leitura
     public ProductDTO findById(Long id){
-        Optional<Product> result = productRepository.findById(id);
-        Product product = result.get();
-        ProductDTO dto = new ProductDTO(product);
-        return dto;
-
-        /*
-        Resumido
-        Product product = productRepository.findById(id).get();
+        Product product = productRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado."));
         return new ProductDTO(product);
-         */
     }
 
     public Page<ProductDTO> findAll(Pageable pageable){
@@ -49,15 +48,28 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = productRepository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = productRepository.save(entity);
-        return new ProductDTO(entity);
+        try{
+            Product entity = productRepository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = productRepository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encotrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try{
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial.");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity ){
